Recipe Book of the Future
We have a Recipe Book. It's out of date. It's unwieldy.

I wanna look at the Recipes for the game in the game.

These scripts allow that to happen.
Running gen_recipes_wiki.exe will update the wiki.txt file in tools with the recipes needed (may need to change where this is placed.)
Running update_guidebook.exe will take wiki.txt and translate it's contents using the RecipesGeneratedHeader.xml into RecipesGenerated.xml.
RecipesGenerated.xml is pointed to by the new Recipes page of the Guidebook.
There is a very solid chance this breaks down. Things that could make it break down:
A recipe for a food with a bizarre special character in its name (think $ or @)
Requiring something that is not a reagent or a solid in a recipe. (It wouldn't break badly from this, it just wouldn't display the ingredients.)
Future formatting changes. (if someone changes Reagents to Liquids, for example.)

However, as-is, I've done just about everything I can to future-proof these scripts, and they're designed and commented in such a way that another person should theoretically be able to come in after me and change the variables at the head of each file to get what they want. If they want to do more than that with them, though... Good luck.